contextbasedbot/chat/serializers.py
from rest_framework import serializers
from .models import Document

class DocumentSerializer(serializers.ModelSerializer):
    class Meta:
        model = Document
        fields = ['doc_id', 'title', 'doc_type', 'added_at']