contextbasedbot/chat/services/web_scraper.py
import requests
from bs4 import BeautifulSoup

def scrap_website(url):
    response = requests.get(url, timeout=10)
    soup = BeautifulSoup(response.text, "html.parser")
    text = soup.get_text(separator=" ", strip=True)
    return text
