pyproject.toml
[tool.poetry]
name = "mentor-assignment2"
version = "0.1.0"
description = ""
authors = ["Mahak Gupta <mahak.gupta@coditas.com>"]
readme = "README.md"

package-mode = false  

[tool.poetry.dependencies]
python = ">=3.10"
django = ">=5.2,<6.0"
djangorestframework = ">=3.16.0,<4.0.0"
djangorestframework-simplejwt = ">=5.5.0,<6.0.0"
pgvector = ">=0.4.1,<0.5.0"
sentence-transformers = ">=4.1.0,<5.0.0"
pypdf2 = ">=3.0.1,<4.0.0"
requests = ">=2.32.3,<3.0.0"
beautifulsoup4 = ">=4.13.4,<5.0.0"
yt-dlp = ">=2025.4.30,<2026.0.0"
torch = ">=2.7.0,<3.0.0"
dotenv = ">=0.9.9,<0.10.0"
psycopg = ">=3.2.7,<4.0.0"
psycopg2 = ">=2.9.10,<3.0.0"
groq = ">=0.24.0,<0.25.0"
pytube = ">=15.0.0,<16.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

